output: Move operation candidate preview rendering
The batch show command builds operation candidates and renders their
preview text from the command entry module. Output modules already own
file-review, hunk, patch, and status prompt presentation.

The project lacks a presentation boundary for operation candidate previews,
so command orchestration sits beside porcelain JSON, terminal headers, and
snippet styling. That makes the show-from entry point carry rendering
details that belong with the other output surfaces.

This commit begins separating candidate preview presentation by moving the
overview and detail renderers into output.candidate_preview. The show-from
command keeps candidate discovery, selection, and state recording while
calling the new output functions.

The final commit will update tests to target the output renderer directly.
